consumer/requirements.txt
Flask==1.1.1
pytest==6.2.5
requests==2.28.0
pact_python==2.0.0